[WIP]: theme/localization resetup
