+++ src/App.tsx
@@ -1,17 +1,17 @@
 import {CrashShield} from '@app/components';
 import React from 'react';
 import {GestureHandlerRootView} from 'react-native-gesture-handler';
-import {AppI18nProvider} from './i18n';
-import CombinedContextProvider from './store/context-providers';
-import PersistedReduxProvider from './store/redux';
-import {AppThemeProvider} from './theme';
-import Navigator from './navigation/Navigator';
+import CombinedContextProvider from '@app/store/context-providers';
+import PersistedReduxProvider from '@app/store/redux';
+import {AppThemeProvider} from '@app/theme';
+import Navigator from '@app/navigation/Navigator';
+import {AppLocalizationProvider} from '@app/i18n';
 
 const App = () => {
   return (
     <CrashShield>
       <GestureHandlerRootView style={{flex: 1}}>
-        <AppI18nProvider>
+        <AppLocalizationProvider>
           <CombinedContextProvider>
             <PersistedReduxProvider>
               <AppThemeProvider autoDetect>
@@ -19,7 +19,7 @@ const App = () => {
               </AppThemeProvider>
             </PersistedReduxProvider>
           </CombinedContextProvider>
-        </AppI18nProvider>
+        </AppLocalizationProvider>
       </GestureHandlerRootView>
     </CrashShield>
   );
